movi favorites actions a users, correccion search encuentra luego de lenght 2
