+++ srcFrontOne/src/Redux/Reducers/flights.js
@@ -14,8 +14,6 @@ import {
     CLEAR_TICKETS,
     POST_TICKET,
     FILTER_PRICE,
-    SET_FAVORITES,
-    DELETE_FAVORITES
 } from "../Constants/flights";
 
 const initialState = {
@@ -26,7 +24,6 @@ const initialState = {
     backup: [],
     flightsByRoute: {},
     getCities: [],
-    favorites: [],
     isFetching: false,  //para controlar si se esta realizando el fetch a la api
     error: false //para cuando no se esta realizando el fetch a la api
 }
@@ -129,17 +126,6 @@ export default flightsReducers = (state = initialState, action) => {
             return {
                 ...state,
                 cart: [],
-            }
-        case SET_FAVORITES:
-            return {
-                ...state,
-                favorites: [action.payload, ...state.favorites]
-            }
-        case DELETE_FAVORITES:
-            const filterFav = state.favorites.filter((flight) => flight.flyId !== action.payload)
-            return {
-                ...state,
-                favorites: filterFav
             }
         case FILTER_PRICE:
             let flagFilt = 0
